+++ gameserver/realm/entity/gameobject/GameObject.Sign.cs
@@ -2,7 +2,7 @@
 {
     partial class Sign : GameObject
     {
-        public Sign(ushort objType)
+        public Sign(int objType)
             : base(objType, null, true, false, false) { }
 
         public override bool HitByProjectile(Projectile projectile, RealmTime time) => false;
